Add the deactivated WARSTATUS code
Sadly, Discord limit the number of publications that you can do in an
hour to 10 by announcement channel, which is too few for such usage. As
such you cannot insert it in your own channels.

It's out of my server scope, but can be used to display the warstatus on other
messaging system.